requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pandas
openpyxl
gunicorn